[ADD]: added doctor store logic and validatioin
