+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Api\DoctorController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -7,3 +8,9 @@ Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
     return $request->user();
 });
 require __DIR__ . '/auth.php';
+
+//Doctor routes
+
+Route::apiResource('doctors', DoctorController::class, )->middleware('auth:sanctum');
+
+//patient api routes
